+++ larva/src/main/java/nl/nn/adapterframework/testtool/LarvaServlet.java
@@ -38,7 +38,7 @@ import nl.nn.adapterframework.util.StreamUtil;
 @IbisInitializer
 public class LarvaServlet extends HttpServletBase {
 	private static final URL INDEX_TEMPLATE = getResource("/index.html.template");
-	private static final String SERVLET_PATH = "/larva/";
+	private static final String SERVLET_PATH = "/iaf/larva/";
 	private final transient Logger log = LogUtil.getLogger(this);
 
 	private enum Assets {
@@ -138,8 +138,7 @@ public class LarvaServlet extends HttpServletBase {
 		resp.setContentType("text/html");
 		writer.append(getTemplate("Larva Test Tool"));
 
-		String cleanServletPath = getUrlMapping().replace("/*", "");
-		String realPath = getServletContext().getRealPath(cleanServletPath);
+		String realPath = getServletContext().getRealPath("/iaf/");
 
 		TestTool.runScenarios(getServletContext(), req, writer, realPath);
 		writer.append("</body></html>");