+++ RELEASES.md
@@ -19,7 +19,7 @@ Upcoming (7.9)
 - The ZipWriterSender with action=WRITE (the default) and no content parameter does no longer have its input as its response, but rather a null message.
 - Parameter with an attribute value set to an empty string will have the empty string as result. Previously the input message would be used. This behaviour can be reobtained by setting
   defaultValueMethod="input".
-
+- Larva context has changed from '<rootcontext>/larva' to '<rootcontext>/iaf/larva'. 
 
 
 7.8-RC1